Extract small modules where possible
