+++ ambiata-mafia.cabal
@@ -43,10 +43,17 @@ library
                     BuildInfo_ambiata_mafia
                     Mafia
                     Mafia.Cabal
+                    Mafia.Cache
+                    Mafia.Error
                     Mafia.Git
+                    Mafia.Home
+                    Mafia.Hoogle
                     Mafia.IO
                     Mafia.Path
                     Mafia.Process
+                    Mafia.Project
+                    Mafia.Sandbox
+                    Mafia.Submodule
 
 executable mafia
   ghc-options:
@@ -62,12 +69,9 @@ executable mafia
                     , ambiata-x-eithert
                     , ambiata-x-optparse
                     , containers                      == 0.5.*
-                    , cryptohash
                     , directory                       == 1.2.*
-                    , exceptions                      == 0.8.*
                     , filepath                        == 1.4.*
                     , optparse-applicative            == 0.11.*
-                    , temporary
                     , text                            == 1.2.*
                     , time                            >= 1.4        && < 1.6
                     , time-locale-compat              == 0.1.*
